Finishing it up
In this update, we have introduced user icons, providing users with the flexibility to change their icons whenever they edit their profiles. These icons have also been seamlessly integrated with the comment section, enhancing the overall user experience. Additionally, we have further developed and optimized the server-side functions for handling contacts.

As part of streamlining the user interface, we have decided to remove the category section since it appeared to be unnecessary for our specific use case. If required, administrative functions can still be accessed through the admin section. This simplification aims to create a more focused and user-friendly environment.